add &mod_tab_chord to solve Alt+Tab race condition
Overlay Alt+tab from Cursor layer onto Number layer to remove the need
for the Cursor layer key to be pressed first before the Number layer key
when triggering the Alt+Tab shortcut.  Note that the Cursor layer must
come below the Number layer in the tower for `&mo LAYER_Cursor` to work.